Baseline of new version of Nextcloud, ready for modification and debugging
Former-commit-id: 31796c2a10e5ec44e793ec7aa8306e31d8786428

+++ nextcloud/lib/private/Template/SCSSCacher.php
@@ -2,12 +2,14 @@
 /**
  * @copyright Copyright (c) 2016, John Molakvoæ (skjnldsv@protonmail.com)
  *
+ * @author Christoph Wurst <christoph@winzerhof-wurst.at>
  * @author John Molakvoæ (skjnldsv) <skjnldsv@protonmail.com>
  * @author Julius Haertl <jus@bitgrid.net>
  * @author Julius Härtl <jus@bitgrid.net>
- * @author Lukas Reschke <lukas@statuscode.ch>
  * @author Morris Jobke <hey@morrisjobke.de>
+ * @author Robin Appelman <robin@icewind.nl>
  * @author Roeland Jago Douma <roeland@famdouma.nl>
+ * @author Roland Tapken <roland@bitarbeiter.net>
  *
  * @license GNU AGPL version 3 or any later version
  *
@@ -22,16 +24,15 @@
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
 namespace OC\Template;
 
-use Leafo\ScssPhp\Compiler;
-use Leafo\ScssPhp\Exception\ParserException;
-use Leafo\ScssPhp\Formatter\Crunched;
-use Leafo\ScssPhp\Formatter\Expanded;
+use OC\AppConfig;
+use OC\Files\AppData\Factory;
+use OC\Memcache\NullCache;
 use OCP\AppFramework\Utility\ITimeFactory;
 use OCP\Files\IAppData;
 use OCP\Files\NotFoundException;
@@ -42,9 +43,11 @@ use OCP\ICache;
 use OCP\ICacheFactory;
 use OCP\IConfig;
 use OCP\ILogger;
+use OCP\IMemcache;
 use OCP\IURLGenerator;
-use OC\Files\AppData\Factory;
-use OC\Template\IconsCacher;
+use ScssPhp\ScssPhp\Compiler;
+use ScssPhp\ScssPhp\Exception\ParserException;
+use ScssPhp\ScssPhp\OutputStyle;
 
 class SCSSCacher {
 
@@ -84,6 +87,11 @@ class SCSSCacher {
 	/** @var ITimeFactory */
 	private $timeFactory;
 
+	/** @var IMemcache */
+	private $lockingCache;
+	/** @var AppConfig */
+	private $appConfig;
+
 	/**
 	 * @param ILogger $logger
 	 * @param Factory $appDataFactory
@@ -103,18 +111,25 @@ class SCSSCacher {
 								$serverRoot,
 								ICacheFactory $cacheFactory,
 								IconsCacher $iconsCacher,
-								ITimeFactory $timeFactory) {
-		$this->logger       = $logger;
-		$this->appData      = $appDataFactory->get('css');
+								ITimeFactory $timeFactory,
+								AppConfig $appConfig) {
+		$this->logger = $logger;
+		$this->appData = $appDataFactory->get('css');
 		$this->urlGenerator = $urlGenerator;
-		$this->config       = $config;
-		$this->defaults     = $defaults;
-		$this->serverRoot   = $serverRoot;
+		$this->config = $config;
+		$this->defaults = $defaults;
+		$this->serverRoot = $serverRoot;
 		$this->cacheFactory = $cacheFactory;
-		$this->depsCache    = $cacheFactory->createDistributed('SCSS-' . md5($this->urlGenerator->getBaseUrl()));
-		$this->isCachedCache = $cacheFactory->createLocal('SCSS-cached-' . md5($this->urlGenerator->getBaseUrl()));
+		$this->depsCache = $cacheFactory->createDistributed('SCSS-deps-' . md5($this->urlGenerator->getBaseUrl()));
+		$this->isCachedCache = $cacheFactory->createDistributed('SCSS-cached-' . md5($this->urlGenerator->getBaseUrl()));
+		$lockingCache = $cacheFactory->createDistributed('SCSS-locks-' . md5($this->urlGenerator->getBaseUrl()));
+		if (!($lockingCache instanceof IMemcache)) {
+			$lockingCache = new NullCache();
+		}
+		$this->lockingCache = $lockingCache;
 		$this->iconsCacher = $iconsCacher;
 		$this->timeFactory = $timeFactory;
+		$this->appConfig = $appConfig;
 	}
 
 	/**
@@ -130,17 +145,15 @@ class SCSSCacher {
 		$path = explode('/', $root . '/' . $file);
 
 		$fileNameSCSS = array_pop($path);
-		$fileNameCSS  = $this->prependVersionPrefix($this->prependBaseurlPrefix(str_replace('.scss', '.css', $fileNameSCSS)), $app);
+		$fileNameCSS = $this->prependVersionPrefix($this->prependBaseurlPrefix(str_replace('.scss', '.css', $fileNameSCSS)), $app);
 
-		$path   = implode('/', $path);
+		$path = implode('/', $path);
 		$webDir = $this->getWebDir($path, $app, $this->serverRoot, \OC::$WEBROOT);
 
+		$this->logger->debug('SCSSCacher::process ordinary check follows', ['app' => 'scss_cacher']);
 		if (!$this->variablesChanged() && $this->isCached($fileNameCSS, $app)) {
 			// Inject icons vars css if any
-			if ($this->iconsCacher->getCachedCSS() && $this->iconsCacher->getCachedCSS()->getSize() > 0) {
-				$this->iconsCacher->injectCss();
-			}
-			return true;
+			return $this->injectCssVariablesIfAny();
 		}
 
 		try {
@@ -150,7 +163,38 @@ class SCSSCacher {
 			$folder = $this->appData->newFolder($app);
 		}
 
-		$cached = $this->cache($path, $fileNameCSS, $fileNameSCSS, $folder, $webDir);
+		$lockKey = $webDir . '/' . $fileNameSCSS;
+
+		if (!$this->lockingCache->add($lockKey, 'locked!', 120)) {
+			$this->logger->debug('SCSSCacher::process could not get lock for ' . $lockKey . ' and will wait 10 seconds for cached file to be available', ['app' => 'scss_cacher']);
+			$retry = 0;
+			sleep(1);
+			while ($retry < 10) {
+				$this->appConfig->clearCachedConfig();
+				$this->logger->debug('SCSSCacher::process check in while loop follows', ['app' => 'scss_cacher']);
+				if (!$this->variablesChanged() && $this->isCached($fileNameCSS, $app)) {
+					// Inject icons vars css if any
+					$this->logger->debug("SCSSCacher::process cached file for app '$app' and file '$fileNameCSS' is now available after $retry s. Moving on...", ['app' => 'scss_cacher']);
+					return $this->injectCssVariablesIfAny();
+				}
+				sleep(1);
+				$retry++;
+			}
+			$this->logger->debug('SCSSCacher::process Giving up scss caching for ' . $lockKey, ['app' => 'scss_cacher']);
+			return false;
+		}
+
+		$this->logger->debug('SCSSCacher::process Lock acquired for ' . $lockKey, ['app' => 'scss_cacher']);
+		try {
+			$cached = $this->cache($path, $fileNameCSS, $fileNameSCSS, $folder, $webDir);
+		} catch (\Exception $e) {
+			$this->lockingCache->remove($lockKey);
+			throw $e;
+		}
+
+		// Cleaning lock
+		$this->lockingCache->remove($lockKey);
+		$this->logger->debug('SCSSCacher::process Lock removed for ' . $lockKey, ['app' => 'scss_cacher']);
 
 		// Inject icons vars css if any
 		if ($this->iconsCacher->getCachedCSS() && $this->iconsCacher->getCachedCSS()->getSize() > 0) {
@@ -166,7 +210,7 @@ class SCSSCacher {
 	 * @return ISimpleFile
 	 */
 	public function getCachedCSS(string $appName, string $fileName): ISimpleFile {
-		$folder         = $this->appData->getFolder($appName);
+		$folder = $this->appData->getFolder($appName);
 		$cachedFileName = $this->prependVersionPrefix($this->prependBaseurlPrefix($fileName), $appName);
 
 		return $folder->getFile($cachedFileName);
@@ -180,44 +224,55 @@ class SCSSCacher {
 	 */
 	private function isCached(string $fileNameCSS, string $app) {
 		$key = $this->config->getSystemValue('version') . '/' . $app . '/' . $fileNameCSS;
-		if (!$this->config->getSystemValue('debug') && $cacheValue = $this->isCachedCache->get($key)) {
+
+		// If the file mtime is more recent than our cached one,
+		// let's consider the file is properly cached
+		if ($cacheValue = $this->isCachedCache->get($key)) {
 			if ($cacheValue > $this->timeFactory->getTime()) {
 				return true;
 			}
 		}
+		$this->logger->debug("SCSSCacher::isCached $fileNameCSS isCachedCache is expired or unset", ['app' => 'scss_cacher']);
 
+		// Creating file cache if none for further checks
 		try {
 			$folder = $this->appData->getFolder($app);
 		} catch (NotFoundException $e) {
-			// creating css appdata folder
-			$folder = $this->appData->newFolder($app);
+			$this->logger->debug("SCSSCacher::isCached app data folder for $app could not be fetched", ['app' => 'scss_cacher']);
+			return false;
 		}
 
+		// Checking if file size is coherent
+		// and if one of the css dependency changed
 		try {
 			$cachedFile = $folder->getFile($fileNameCSS);
 			if ($cachedFile->getSize() > 0) {
 				$depFileName = $fileNameCSS . '.deps';
-				$deps        = $this->depsCache->get($folder->getName() . '-' . $depFileName);
+				$deps = $this->depsCache->get($folder->getName() . '-' . $depFileName);
 				if ($deps === null) {
 					$depFile = $folder->getFile($depFileName);
-					$deps    = $depFile->getContent();
-					//Set to memcache for next run
+					$deps = $depFile->getContent();
+					// Set to memcache for next run
 					$this->depsCache->set($folder->getName() . '-' . $depFileName, $deps);
 				}
 				$deps = json_decode($deps, true);
 
 				foreach ((array) $deps as $file => $mtime) {
 					if (!file_exists($file) || filemtime($file) > $mtime) {
+						$this->logger->debug("SCSSCacher::isCached $fileNameCSS is not considered as cached due to deps file $file", ['app' => 'scss_cacher']);
 						return false;
 					}
 				}
 
+				$this->logger->debug("SCSSCacher::isCached $fileNameCSS dependencies successfully cached for 5 minutes", ['app' => 'scss_cacher']);
+				// It would probably make sense to adjust this timeout to something higher and see if that has some effect then
 				$this->isCachedCache->set($key, $this->timeFactory->getTime() + 5 * 60);
 				return true;
 			}
-
+			$this->logger->debug("SCSSCacher::isCached $fileNameCSS is not considered as cached cacheValue: $cacheValue", ['app' => 'scss_cacher']);
 			return false;
 		} catch (NotFoundException $e) {
+			$this->logger->debug("SCSSCacher::isCached NotFoundException " . $e->getMessage(), ['app' => 'scss_cacher']);
 			return false;
 		}
 	}
@@ -227,14 +282,14 @@ class SCSSCacher {
 	 * @return bool
 	 */
 	private function variablesChanged(): bool {
-		$injectedVariables = $this->getInjectedVariables();
-		if ($this->config->getAppValue('core', 'scss.variables') !== md5($injectedVariables)) {
+		$cachedVariables = $this->config->getAppValue('core', 'theming.variables', '');
+		$injectedVariables = $this->getInjectedVariables($cachedVariables);
+		if ($cachedVariables !== md5($injectedVariables)) {
+			$this->logger->debug('SCSSCacher::variablesChanged storedVariables: ' . json_encode($this->config->getAppValue('core', 'theming.variables')) . ' currentInjectedVariables: ' . json_encode($injectedVariables), ['app' => 'scss_cacher']);
+			$this->config->setAppValue('core', 'theming.variables', md5($injectedVariables));
 			$this->resetCache();
-			$this->config->setAppValue('core', 'scss.variables', md5($injectedVariables));
-
 			return true;
 		}
-
 		return false;
 	}
 
@@ -257,14 +312,12 @@ class SCSSCacher {
 		]);
 
 		// Continue after throw
-		$scss->setIgnoreErrors(true);
 		if ($this->config->getSystemValue('debug')) {
 			// Debug mode
-			$scss->setFormatter(Expanded::class);
-			$scss->setLineNumberStyle(Compiler::LINE_COMMENTS);
+			$scss->setOutputStyle(OutputStyle::EXPANDED);
 		} else {
 			// Compression
-			$scss->setFormatter(Crunched::class);
+			$scss->setOutputStyle(OutputStyle::COMPRESSED);
 		}
 
 		try {
@@ -289,7 +342,7 @@ class SCSSCacher {
 				'@import "functions.scss";' .
 				'@import "' . $fileNameSCSS . '";');
 		} catch (ParserException $e) {
-			$this->logger->error($e, ['app' => 'core']);
+			$this->logger->logException($e, ['app' => 'scss_cacher']);
 
 			return false;
 		}
@@ -311,11 +364,11 @@ class SCSSCacher {
 			$depFile->putContent($deps);
 			$this->depsCache->set($folder->getName() . '-' . $depFileName, $deps);
 			$gzipFile->putContent(gzencode($data, 9));
-			$this->logger->debug('SCSSCacher: ' . $webDir . '/' . $fileNameSCSS . ' compiled and successfully cached', ['app' => 'core']);
+			$this->logger->debug('SCSSCacher::cache ' . $webDir . '/' . $fileNameSCSS . ' compiled and successfully cached', ['app' => 'scss_cacher']);
 
 			return true;
 		} catch (NotPermittedException $e) {
-			$this->logger->error('SCSSCacher: unable to cache: ' . $fileNameSCSS);
+			$this->logger->error('SCSSCacher::cache unable to cache: ' . $fileNameSCSS, ['app' => 'scss_cacher']);
 
 			return false;
 		}
@@ -326,24 +379,37 @@ class SCSSCacher {
 	 * We need to regenerate all files when variables change
 	 */
 	public function resetCache() {
+		$this->logger->debug('SCSSCacher::resetCache', ['app' => 'scss_cacher']);
+		if (!$this->lockingCache->add('resetCache', 'locked!', 120)) {
+			$this->logger->debug('SCSSCacher::resetCache Locked', ['app' => 'scss_cacher']);
+			return;
+		}
+		$this->logger->debug('SCSSCacher::resetCache Lock acquired', ['app' => 'scss_cacher']);
 		$this->injectedVariables = null;
-		$this->cacheFactory->createDistributed('SCSS-')->clear();
+
+		// do not clear locks
+		$this->depsCache->clear();
+		$this->isCachedCache->clear();
+
 		$appDirectory = $this->appData->getDirectoryListing();
 		foreach ($appDirectory as $folder) {
 			foreach ($folder->getDirectoryListing() as $file) {
 				try {
 					$file->delete();
 				} catch (NotPermittedException $e) {
-					$this->logger->logException($e, ['message' => 'SCSSCacher: unable to delete file: ' . $file->getName()]);
+					$this->logger->logException($e, ['message' => 'SCSSCacher::resetCache unable to delete file: ' . $file->getName(), 'app' => 'scss_cacher']);
 				}
 			}
 		}
+		$this->logger->debug('SCSSCacher::resetCache css cache cleared!', ['app' => 'scss_cacher']);
+		$this->lockingCache->remove('resetCache');
+		$this->logger->debug('SCSSCacher::resetCache Locking removed', ['app' => 'scss_cacher']);
 	}
 
 	/**
 	 * @return string SCSS code for variables from OC_Defaults
 	 */
-	private function getInjectedVariables(): string {
+	private function getInjectedVariables(string $cache = ''): string {
 		if ($this->injectedVariables !== null) {
 			return $this->injectedVariables;
 		}
@@ -352,13 +418,22 @@ class SCSSCacher {
 			$variables .= '$' . $key . ': ' . $value . ' !default;';
 		}
 
+		/*
+		 * If we are trying to return the same variables as that are cached
+		 * Then there is no need to do the compile step
+		 */
+		if ($cache === md5($variables)) {
+			$this->injectedVariables = $variables;
+			return $variables;
+		}
+
 		// check for valid variables / otherwise fall back to defaults
 		try {
 			$scss = new Compiler();
 			$scss->compile($variables);
 			$this->injectedVariables = $variables;
 		} catch (ParserException $e) {
-			$this->logger->error($e, ['app' => 'core']);
+			$this->logger->logException($e, ['app' => 'scss_cacher']);
 		}
 
 		return $variables;
@@ -371,7 +446,7 @@ class SCSSCacher {
 	 * @return string
 	 */
 	private function rebaseUrls(string $css, string $webDir): string {
-		$re    = '/url\([\'"]([^\/][\.\w?=\/-]*)[\'"]\)/x';
+		$re = '/url\([\'"]([^\/][\.\w?=\/-]*)[\'"]\)/x';
 		$subst = 'url(\'' . $webDir . '/$1\')';
 
 		return preg_replace($re, $subst, $css);
@@ -385,13 +460,13 @@ class SCSSCacher {
 	 */
 	public function getCachedSCSS(string $appName, string $fileName): string {
 		$tmpfileLoc = explode('/', $fileName);
-		$fileName   = array_pop($tmpfileLoc);
-		$fileName   = $this->prependVersionPrefix($this->prependBaseurlPrefix(str_replace('.scss', '.css', $fileName)), $appName);
+		$fileName = array_pop($tmpfileLoc);
+		$fileName = $this->prependVersionPrefix($this->prependBaseurlPrefix(str_replace('.scss', '.css', $fileName)), $appName);
 
 		return substr($this->urlGenerator->linkToRoute('core.Css.getCss', [
 			'fileName' => $fileName,
 			'appName' => $appName,
-			'v' => $this->config->getAppValue('core', 'scss.variables', '0')
+			'v' => $this->config->getAppValue('core', 'theming.variables', '0')
 		]), \strlen(\OC::$WEBROOT) + 1);
 	}
 
@@ -448,5 +523,18 @@ class SCSSCacher {
 		}
 
 		return $webRoot . substr($path, strlen($serverRoot));
+	}
+
+	/**
+	 * Add the icons css cache in the header if needed
+	 *
+	 * @return boolean true
+	 */
+	private function injectCssVariablesIfAny() {
+		// Inject icons vars css if any
+		if ($this->iconsCacher->getCachedCSS() && $this->iconsCacher->getCachedCSS()->getSize() > 0) {
+			$this->iconsCacher->injectCss();
+		}
+		return true;
 	}
 }
